qemu: enter interruptlible state where appropriate
This is just places where async job drops vm lock besides
for the purpose of sending command to qemu process. If I fail
to identify one of these places it is not a big problem - vm will
be unresponsive to queries etc sometimes and someone who cares
report this issue and we fix it ))

NB: Now there are some known places were enter/exit are not yet added
as noted in the cover letter. These are end/begin of migration phases.
Message-Id: <1510233537-848259-4-git-send-email-nshirokovskiy@virtuozzo.com>