qemu: remove nested job usage from qemuProcessStop
81f50cb92 added getting nested job on stopping domain to prevent leaking of
monitor object on race between aborting job and stopping domain cased by that
abort.

One of the causes of this problem is that async job and concurrent regular job
were not fully isolated and async job can continue to run when concurrent job
is not finished yet. Now this is not possible and we can safely remove getting
nested job condition in qemuProcessStop as stopping always done within context
of some job.
Message-Id: <1510233537-848259-5-git-send-email-nshirokovskiy@virtuozzo.com>