qemu: use job to keep driver responsive on qemuDomainOpenChannel
As we now assume that virChrdevOpen can take a significant time
we should take a job and drop vm lock during operation to keep
qemu driver responsive.

We drop vm lock but access vm structures after it because we are
in exlusive modify job so vm structures can only be read concurrently.
Message-Id: <1510844672-47857-4-git-send-email-nshirokovskiy@virtuozzo.com>