qemu: add support for setting OEM strings SMBIOS data fields
The OEM strings table in SMBIOS allows the vendor to pass arbitrary strings
into the guest OS. This can be used as a way to pass data to an application like
cloud-init, or potentially as an alternative to the kernel command line for OS
installers where you can't modify the install ISO image to change the kernel
args.

As an example, consider if cloud-init and anaconda supported OEM strings you
could use something like

    <oemStrings>
      <entry>cloud-init:ds=nocloud-net;s=http://10.10.0.1:8000/</entry>
      <entry>anaconda:method=http://dl.fedoraproject.org/pub/fedora/linux/releases/25/x86_64/os</entry>
    </oemStrings>

use of a application specific prefix as illustrated above is recommended so that
an app can reliably identify which of the many OEM strings are targetted at it.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>
Message-Id: <20171117164853.29166-1-berrange@redhat.com>