qemu: Allow showing the dump progress for memory only dump
https://bugzilla.redhat.com/show_bug.cgi?id=916061

If the QEMU version running is new enough (based on QEMU_CAPS_QUERY_DUMP)
we can add a 'detach' boolean to the dump-guest-memory command in order
to tell QEMU to run in a thread. Then, use the qemuDumpWaitForCompletion
in order to 'watch' the dump and save the stats into the jobInfo data
so that tools (such as virsh) can pick up and display the completion
percentage. The processing is similar to qemuMigrationWaitForCompletion
at least with respect to calling the query-dump in a while loop that
gets a micro sleep in order for qemuDomainGetJobInfo to be able to
pull out the changed migration stats values.

As an added benefit, while the domain is being dumped, we don't lock
out other commands.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20171117231731.2880-7-jferlan@redhat.com>