qemu: command: Split out geometry frontend attribute formatting from -drive
Historically we've formatted a lot of the attributes of a disk (disk
geometry, etc) with -drive. Since we use -device now, they should be
formatted there. Extract them to a separate function for keeping
compatibility with SDcards which still use only -drive.

Start this by moving the geometry into a separate function.
Message-Id: <683f8bae847f8bc3510185b9c6f8db6cd8b9680e.1511198532.git.pkrempa@redhat.com>