qemu: Introduce qemuProcessHandleDumpCompleted
Add a couple of booleans to mark when there's a dump completion event
waiting and when a dump completed event has been received.

To ensure the dump completed event from a non memory-only dump doesn't
cause the a dump completed event to be fired, only broadcast if there's
a completion event waiting.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20171120212121.7565-5-jferlan@redhat.com>