conf: Shorten names in virDomainChrSerialTarget enumeration
Now that the target type is no longer formatted on the QEMU command
line, we don't need the values to match the QEMU device names any
longer, so we can shorten the names and reduce redundancy by dropping
the -serial suffix: this also has the nice side-effect that target
type and address type will now match.

We still need to parse the old names, and format them when preparing
a migratable XML, to preserve backwards compatibility.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20171121164231.15008-15-abologna@redhat.com>