qemu: command: Inject password-secret only when not using JSON props
The 'file.password-secret' injection should be used only if we are using
the old formatter. When formatting the source string from the JSON
properities, the property should be added there.

Also drop the comment which refers to stuff that will not be used in
libvirt sine -blockdev is the way to go.
Message-Id: <f11a3c96e37cb0412087f8f5a6430af3d02a8fe0.1511454145.git.pkrempa@redhat.com>