qemu: domain: Fix backing store terminator for non-backing local files
Raw local files do not pass through the backing store detector and thus
the code did not allocate the required backing store terminator for
them. Previously the terminating element would be formatted into the XML
since the default values used for the metadata allowed that. This is a
regression since a693fdba0111ff which was not detected in the review.

This patch also reverts all the changes in the test files.
Message-Id: <d9ecb74f9e7f822f139fb4820d33800806e6c1c9.1511526015.git.pkrempa@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>