virfile: relax checks for hugetlbfs
There are use cases where it is useful to use the support in libvirt
for file-backed guest memory, but without using hugetlbfs but tmpfs
instead (for example, to run tests on hosts that do not have hugepages
configured, or to use Linux's idle page tracking to monitor guest
memory accesses at a 4k page granularity.).

Drop the check for hugetlbfs when querying the huge page size, but
move it to the loop that's searching for a suitable mount to use.

Change-Id: I2c9589191e14653724725b944171689553ee6bae
Message-Id: <20171204152127.14534-1-david.vrabel@nutanix.com>