qemu: Introduce qemuDomainDeviceDefSkipController
In order to be able to reuse some code for both controller def
validation and command line building, create a convenience helper
that will perform the "skip" avoidance checks. It will also set
some flags to allow the caller to specifically skip (or fail)
depending on the result of the flag (as opposed to building up
some ever growing list of variables).

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20171205013907.8563-3-jferlan@redhat.com>