qemu: Complete move USB command checks to controller def validate
Move the remainder of the qemuBuildUSBControllerDevStr checks over
to qemuDomainDeviceDefValidateControllerUSB. This also allows the
command code to shorten up a bit and become a void procedure.

This also requires modifying the xml2xml test a bit in order to
include the correct capability bit and modifying the xml2argv test
in order to use the PARSE_ERROR macro.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20171205013907.8563-18-jferlan@redhat.com>