qemu: Complete PCI command checks to controller def validate
Move the qemuCaps checks over to qemuDomainControllerDefValidatePCI
and add empty lines between cases so it's a bit easier on the eyes
to read.

This requires two test updates in order to set the correct capability
bit for an xml2xml test as well as setting up the similar capability
for the pseries memlocktest.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20171207011408.15091-12-jferlan@redhat.com>