nodedev: Move the udevPCITranslateInit call
If the timing is "just right", there is a possibility that the
udev nodeStateInitialize conflicts with another systemd thread
running an lspci command leaving both waiting for "something",
but resulting in a hung libvirtd (and hung lspci thread) from
which the only recovery is a reboot because killing either thread
is impossible and results in a defunct libvirtd process if a
SIGKILL is performed.

In order to avoid this let's move where the PCI initialization
is done to be where it's actually needed. Ensure we only perform
the initialization once via a driver bool.  Likewise, during
cleanup ensure we only call udevPCITranslateDeinit once the
initialization is successful.

At least a failure for this driver won't hang out the rest of the
the libvirt event loop. May not make certain things usable though.
Still a libvirtd restart is far easier than a host reboot.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20171209172914.14640-4-jferlan@redhat.com>